Pull news item into its own file
